+++ src/main/java/com/back/global/config/WebSocketConfig.java
@@ -1,7 +1,7 @@
 package com.back.global.config;
 
-import com.back.global.security.CustomUserDetails;
-import com.back.global.security.JwtTokenProvider;
+import com.back.global.security.user.CustomUserDetails;
+import com.back.global.security.jwt.JwtTokenProvider;
 import com.back.global.websocket.service.WebSocketSessionManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
