+++ src/main/java/com/back/global/security/jwt/JwtTokenProvider.java
@@ -1,8 +1,9 @@
-package com.back.global.security;
+package com.back.global.security.jwt;
 
 import com.back.domain.user.entity.Role;
 import com.back.global.exception.CustomException;
 import com.back.global.exception.ErrorCode;
+import com.back.global.security.user.CustomUserDetails;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
