+++ src/test/java/com/back/fixture/TestController.java
@@ -1,6 +1,6 @@
 package com.back.fixture;
 
-import com.back.global.security.CustomUserDetails;
+import com.back.global.security.user.CustomUserDetails;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
