+++ src/test/java/com/back/global/security/JwtSecurityIntegrationTest.java
@@ -1,5 +1,6 @@
 package com.back.global.security;
 
+import com.back.global.security.jwt.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
